feat: finish add tweet function in Navbar
